Cpp file compatability added
